requirements.txt
Django==1.4.2
South==0.7.6
boto==2.6.0
distribute==0.6.28
dj-database-url==0.2.1
django-dajax==0.9.2
django-dajaxice==0.5.2
django-storages==1.1.5
gunicorn==0.15.0
psycopg2==2.4.5
simplejson==2.6.2
wsgiref==0.1.2
